requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
PyPDF2==3.0.1
requests==2.31.0
pydantic==2.3.0
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1
pdf2image==1.16.3
Pillow==10.0.0 